Created Math namespace and added functions and tests
I created the Math namespace and added two classes: MathComponents and Equation. An equation contains one or many MathComponent, which can represent an operand or a floating point number. I also added the Calculation function as well as the unit tests for each. The Math library is not yet implemented in the project but will be soon.